partition Mimic.Server
This commit changes Mimic.Server from a single GenServer to be partitioned
using a PartitionSupervisor.

Closes #115.
